"Cambio sprites y agrego verificacion del Json"

+++ src/Common.h
@@ -95,7 +95,7 @@ struct Tventana {
 struct Tescenario {
     float ancho;
     float alto;
-    int yPiso;
+    float yPiso;
 };
 
 struct Tcapa {
